fix refactor caching
