+++ litellm/caching/caching.py
@@ -26,9 +26,13 @@ from litellm._logging import verbose_logger
 from litellm.types.caching import *
 from litellm.types.utils import all_litellm_params
 
+from .base_cache import BaseCache
+from .disk_cache import DiskCache
+from .dual_cache import DualCache
 from .in_memory_cache import InMemoryCache
 from .qdrant_semantic_cache import QdrantSemanticCache
 from .redis_cache import RedisCache
+from .redis_semantic_cache import RedisSemanticCache
 from .s3_cache import S3Cache
 
 
@@ -44,663 +48,6 @@ def print_verbose(print_statement):
 class CacheMode(str, Enum):
     default_on = "default_on"
     default_off = "default_off"
-
-
-class BaseCache:
-    def set_cache(self, key, value, **kwargs):
-        raise NotImplementedError
-
-    async def async_set_cache(self, key, value, **kwargs):
-        raise NotImplementedError
-
-    def get_cache(self, key, **kwargs):
-        raise NotImplementedError
-
-    async def async_get_cache(self, key, **kwargs):
-        raise NotImplementedError
-
-    async def batch_cache_write(self, result, *args, **kwargs):
-        raise NotImplementedError
-
-    async def disconnect(self):
-        raise NotImplementedError
-
-
-class RedisSemanticCache(BaseCache):
-    def __init__(
-        self,
-        host=None,
-        port=None,
-        password=None,
-        redis_url=None,
-        similarity_threshold=None,
-        use_async=False,
-        embedding_model="text-embedding-ada-002",
-        **kwargs,
-    ):
-        from redisvl.index import SearchIndex
-        from redisvl.query import VectorQuery
-
-        print_verbose(
-            "redis semantic-cache initializing INDEX - litellm_semantic_cache_index"
-        )
-        if similarity_threshold is None:
-            raise Exception("similarity_threshold must be provided, passed None")
-        self.similarity_threshold = similarity_threshold
-        self.embedding_model = embedding_model
-        schema = {
-            "index": {
-                "name": "litellm_semantic_cache_index",
-                "prefix": "litellm",
-                "storage_type": "hash",
-            },
-            "fields": {
-                "text": [{"name": "response"}],
-                "vector": [
-                    {
-                        "name": "litellm_embedding",
-                        "dims": 1536,
-                        "distance_metric": "cosine",
-                        "algorithm": "flat",
-                        "datatype": "float32",
-                    }
-                ],
-            },
-        }
-        if redis_url is None:
-            # if no url passed, check if host, port and password are passed, if not raise an Exception
-            if host is None or port is None or password is None:
-                # try checking env for host, port and password
-                import os
-
-                host = os.getenv("REDIS_HOST")
-                port = os.getenv("REDIS_PORT")
-                password = os.getenv("REDIS_PASSWORD")
-                if host is None or port is None or password is None:
-                    raise Exception("Redis host, port, and password must be provided")
-
-            redis_url = "redis://:" + password + "@" + host + ":" + port
-        print_verbose(f"redis semantic-cache redis_url: {redis_url}")
-        if use_async is False:
-            self.index = SearchIndex.from_dict(schema)
-            self.index.connect(redis_url=redis_url)
-            try:
-                self.index.create(overwrite=False)  # don't overwrite existing index
-            except Exception as e:
-                print_verbose(f"Got exception creating semantic cache index: {str(e)}")
-        elif use_async is True:
-            schema["index"]["name"] = "litellm_semantic_cache_index_async"
-            self.index = SearchIndex.from_dict(schema)
-            self.index.connect(redis_url=redis_url, use_async=True)
-
-    #
-    def _get_cache_logic(self, cached_response: Any):
-        """
-        Common 'get_cache_logic' across sync + async redis client implementations
-        """
-        if cached_response is None:
-            return cached_response
-
-        # check if cached_response is bytes
-        if isinstance(cached_response, bytes):
-            cached_response = cached_response.decode("utf-8")
-
-        try:
-            cached_response = json.loads(
-                cached_response
-            )  # Convert string to dictionary
-        except Exception:
-            cached_response = ast.literal_eval(cached_response)
-        return cached_response
-
-    def set_cache(self, key, value, **kwargs):
-        import numpy as np
-
-        print_verbose(f"redis semantic-cache set_cache, kwargs: {kwargs}")
-
-        # get the prompt
-        messages = kwargs["messages"]
-        prompt = "".join(message["content"] for message in messages)
-
-        # create an embedding for prompt
-        embedding_response = litellm.embedding(
-            model=self.embedding_model,
-            input=prompt,
-            cache={"no-store": True, "no-cache": True},
-        )
-
-        # get the embedding
-        embedding = embedding_response["data"][0]["embedding"]
-
-        # make the embedding a numpy array, convert to bytes
-        embedding_bytes = np.array(embedding, dtype=np.float32).tobytes()
-        value = str(value)
-        assert isinstance(value, str)
-
-        new_data = [
-            {"response": value, "prompt": prompt, "litellm_embedding": embedding_bytes}
-        ]
-
-        # Add more data
-        self.index.load(new_data)
-
-        return
-
-    def get_cache(self, key, **kwargs):
-        print_verbose(f"sync redis semantic-cache get_cache, kwargs: {kwargs}")
-        import numpy as np
-        from redisvl.query import VectorQuery
-
-        # query
-        # get the messages
-        messages = kwargs["messages"]
-        prompt = "".join(message["content"] for message in messages)
-
-        # convert to embedding
-        embedding_response = litellm.embedding(
-            model=self.embedding_model,
-            input=prompt,
-            cache={"no-store": True, "no-cache": True},
-        )
-
-        # get the embedding
-        embedding = embedding_response["data"][0]["embedding"]
-
-        query = VectorQuery(
-            vector=embedding,
-            vector_field_name="litellm_embedding",
-            return_fields=["response", "prompt", "vector_distance"],
-            num_results=1,
-        )
-
-        results = self.index.query(query)
-        if results is None:
-            return None
-        if isinstance(results, list):
-            if len(results) == 0:
-                return None
-
-        vector_distance = results[0]["vector_distance"]
-        vector_distance = float(vector_distance)
-        similarity = 1 - vector_distance
-        cached_prompt = results[0]["prompt"]
-
-        # check similarity, if more than self.similarity_threshold, return results
-        print_verbose(
-            f"semantic cache: similarity threshold: {self.similarity_threshold}, similarity: {similarity}, prompt: {prompt}, closest_cached_prompt: {cached_prompt}"
-        )
-        if similarity > self.similarity_threshold:
-            # cache hit !
-            cached_value = results[0]["response"]
-            print_verbose(
-                f"got a cache hit, similarity: {similarity}, Current prompt: {prompt}, cached_prompt: {cached_prompt}"
-            )
-            return self._get_cache_logic(cached_response=cached_value)
-        else:
-            # cache miss !
-            return None
-
-        pass
-
-    async def async_set_cache(self, key, value, **kwargs):
-        import numpy as np
-
-        from litellm.proxy.proxy_server import llm_model_list, llm_router
-
-        try:
-            await self.index.acreate(overwrite=False)  # don't overwrite existing index
-        except Exception as e:
-            print_verbose(f"Got exception creating semantic cache index: {str(e)}")
-        print_verbose(f"async redis semantic-cache set_cache, kwargs: {kwargs}")
-
-        # get the prompt
-        messages = kwargs["messages"]
-        prompt = "".join(message["content"] for message in messages)
-        # create an embedding for prompt
-        router_model_names = (
-            [m["model_name"] for m in llm_model_list]
-            if llm_model_list is not None
-            else []
-        )
-        if llm_router is not None and self.embedding_model in router_model_names:
-            user_api_key = kwargs.get("metadata", {}).get("user_api_key", "")
-            embedding_response = await llm_router.aembedding(
-                model=self.embedding_model,
-                input=prompt,
-                cache={"no-store": True, "no-cache": True},
-                metadata={
-                    "user_api_key": user_api_key,
-                    "semantic-cache-embedding": True,
-                    "trace_id": kwargs.get("metadata", {}).get("trace_id", None),
-                },
-            )
-        else:
-            # convert to embedding
-            embedding_response = await litellm.aembedding(
-                model=self.embedding_model,
-                input=prompt,
-                cache={"no-store": True, "no-cache": True},
-            )
-
-        # get the embedding
-        embedding = embedding_response["data"][0]["embedding"]
-
-        # make the embedding a numpy array, convert to bytes
-        embedding_bytes = np.array(embedding, dtype=np.float32).tobytes()
-        value = str(value)
-        assert isinstance(value, str)
-
-        new_data = [
-            {"response": value, "prompt": prompt, "litellm_embedding": embedding_bytes}
-        ]
-
-        # Add more data
-        await self.index.aload(new_data)
-        return
-
-    async def async_get_cache(self, key, **kwargs):
-        print_verbose(f"async redis semantic-cache get_cache, kwargs: {kwargs}")
-        import numpy as np
-        from redisvl.query import VectorQuery
-
-        from litellm.proxy.proxy_server import llm_model_list, llm_router
-
-        # query
-        # get the messages
-        messages = kwargs["messages"]
-        prompt = "".join(message["content"] for message in messages)
-
-        router_model_names = (
-            [m["model_name"] for m in llm_model_list]
-            if llm_model_list is not None
-            else []
-        )
-        if llm_router is not None and self.embedding_model in router_model_names:
-            user_api_key = kwargs.get("metadata", {}).get("user_api_key", "")
-            embedding_response = await llm_router.aembedding(
-                model=self.embedding_model,
-                input=prompt,
-                cache={"no-store": True, "no-cache": True},
-                metadata={
-                    "user_api_key": user_api_key,
-                    "semantic-cache-embedding": True,
-                    "trace_id": kwargs.get("metadata", {}).get("trace_id", None),
-                },
-            )
-        else:
-            # convert to embedding
-            embedding_response = await litellm.aembedding(
-                model=self.embedding_model,
-                input=prompt,
-                cache={"no-store": True, "no-cache": True},
-            )
-
-        # get the embedding
-        embedding = embedding_response["data"][0]["embedding"]
-
-        query = VectorQuery(
-            vector=embedding,
-            vector_field_name="litellm_embedding",
-            return_fields=["response", "prompt", "vector_distance"],
-        )
-        results = await self.index.aquery(query)
-        if results is None:
-            kwargs.setdefault("metadata", {})["semantic-similarity"] = 0.0
-            return None
-        if isinstance(results, list):
-            if len(results) == 0:
-                kwargs.setdefault("metadata", {})["semantic-similarity"] = 0.0
-                return None
-
-        vector_distance = results[0]["vector_distance"]
-        vector_distance = float(vector_distance)
-        similarity = 1 - vector_distance
-        cached_prompt = results[0]["prompt"]
-
-        # check similarity, if more than self.similarity_threshold, return results
-        print_verbose(
-            f"semantic cache: similarity threshold: {self.similarity_threshold}, similarity: {similarity}, prompt: {prompt}, closest_cached_prompt: {cached_prompt}"
-        )
-
-        # update kwargs["metadata"] with similarity, don't rewrite the original metadata
-        kwargs.setdefault("metadata", {})["semantic-similarity"] = similarity
-
-        if similarity > self.similarity_threshold:
-            # cache hit !
-            cached_value = results[0]["response"]
-            print_verbose(
-                f"got a cache hit, similarity: {similarity}, Current prompt: {prompt}, cached_prompt: {cached_prompt}"
-            )
-            return self._get_cache_logic(cached_response=cached_value)
-        else:
-            # cache miss !
-            return None
-        pass
-
-    async def _index_info(self):
-        return await self.index.ainfo()
-
-
-class DualCache(BaseCache):
-    """
-    This updates both Redis and an in-memory cache simultaneously.
-    When data is updated or inserted, it is written to both the in-memory cache + Redis.
-    This ensures that even if Redis hasn't been updated yet, the in-memory cache reflects the most recent data.
-    """
-
-    def __init__(
-        self,
-        in_memory_cache: Optional[InMemoryCache] = None,
-        redis_cache: Optional[RedisCache] = None,
-        default_in_memory_ttl: Optional[float] = None,
-        default_redis_ttl: Optional[float] = None,
-        always_read_redis: Optional[bool] = True,
-    ) -> None:
-        super().__init__()
-        # If in_memory_cache is not provided, use the default InMemoryCache
-        self.in_memory_cache = in_memory_cache or InMemoryCache()
-        # If redis_cache is not provided, use the default RedisCache
-        self.redis_cache = redis_cache
-
-        self.default_in_memory_ttl = (
-            default_in_memory_ttl or litellm.default_in_memory_ttl
-        )
-        self.default_redis_ttl = default_redis_ttl or litellm.default_redis_ttl
-        self.always_read_redis = always_read_redis
-
-    def update_cache_ttl(
-        self, default_in_memory_ttl: Optional[float], default_redis_ttl: Optional[float]
-    ):
-        if default_in_memory_ttl is not None:
-            self.default_in_memory_ttl = default_in_memory_ttl
-
-        if default_redis_ttl is not None:
-            self.default_redis_ttl = default_redis_ttl
-
-    def set_cache(self, key, value, local_only: bool = False, **kwargs):
-        # Update both Redis and in-memory cache
-        try:
-            if self.in_memory_cache is not None:
-                if "ttl" not in kwargs and self.default_in_memory_ttl is not None:
-                    kwargs["ttl"] = self.default_in_memory_ttl
-
-                self.in_memory_cache.set_cache(key, value, **kwargs)
-
-            if self.redis_cache is not None and local_only is False:
-                self.redis_cache.set_cache(key, value, **kwargs)
-        except Exception as e:
-            print_verbose(e)
-
-    def increment_cache(
-        self, key, value: int, local_only: bool = False, **kwargs
-    ) -> int:
-        """
-        Key - the key in cache
-
-        Value - int - the value you want to increment by
-
-        Returns - int - the incremented value
-        """
-        try:
-            result: int = value
-            if self.in_memory_cache is not None:
-                result = self.in_memory_cache.increment_cache(key, value, **kwargs)
-
-            if self.redis_cache is not None and local_only is False:
-                result = self.redis_cache.increment_cache(key, value, **kwargs)
-
-            return result
-        except Exception as e:
-            verbose_logger.error(f"LiteLLM Cache: Excepton async add_cache: {str(e)}")
-            raise e
-
-    def get_cache(self, key, local_only: bool = False, **kwargs):
-        # Try to fetch from in-memory cache first
-        try:
-            result = None
-            if self.in_memory_cache is not None:
-                in_memory_result = self.in_memory_cache.get_cache(key, **kwargs)
-
-                if in_memory_result is not None:
-                    result = in_memory_result
-
-            if (
-                (self.always_read_redis is True)
-                and self.redis_cache is not None
-                and local_only is False
-            ):
-                # If not found in in-memory cache or always_read_redis is True, try fetching from Redis
-                redis_result = self.redis_cache.get_cache(key, **kwargs)
-
-                if redis_result is not None:
-                    # Update in-memory cache with the value from Redis
-                    self.in_memory_cache.set_cache(key, redis_result, **kwargs)
-
-                result = redis_result
-
-            print_verbose(f"get cache: cache result: {result}")
-            return result
-        except Exception:
-            verbose_logger.error(traceback.format_exc())
-
-    def batch_get_cache(self, keys: list, local_only: bool = False, **kwargs):
-        try:
-            result = [None for _ in range(len(keys))]
-            if self.in_memory_cache is not None:
-                in_memory_result = self.in_memory_cache.batch_get_cache(keys, **kwargs)
-
-                if in_memory_result is not None:
-                    result = in_memory_result
-
-            if None in result and self.redis_cache is not None and local_only is False:
-                """
-                - for the none values in the result
-                - check the redis cache
-                """
-                sublist_keys = [
-                    key for key, value in zip(keys, result) if value is None
-                ]
-                # If not found in in-memory cache, try fetching from Redis
-                redis_result = self.redis_cache.batch_get_cache(sublist_keys, **kwargs)
-                if redis_result is not None:
-                    # Update in-memory cache with the value from Redis
-                    for key in redis_result:
-                        self.in_memory_cache.set_cache(key, redis_result[key], **kwargs)
-
-                for key, value in redis_result.items():
-                    result[keys.index(key)] = value
-
-            print_verbose(f"async batch get cache: cache result: {result}")
-            return result
-        except Exception:
-            verbose_logger.error(traceback.format_exc())
-
-    async def async_get_cache(self, key, local_only: bool = False, **kwargs):
-        # Try to fetch from in-memory cache first
-        try:
-            print_verbose(
-                f"async get cache: cache key: {key}; local_only: {local_only}"
-            )
-            result = None
-            if self.in_memory_cache is not None:
-                in_memory_result = await self.in_memory_cache.async_get_cache(
-                    key, **kwargs
-                )
-
-                print_verbose(f"in_memory_result: {in_memory_result}")
-                if in_memory_result is not None:
-                    result = in_memory_result
-
-            if result is None and self.redis_cache is not None and local_only is False:
-                # If not found in in-memory cache, try fetching from Redis
-                redis_result = await self.redis_cache.async_get_cache(key, **kwargs)
-
-                if redis_result is not None:
-                    # Update in-memory cache with the value from Redis
-                    await self.in_memory_cache.async_set_cache(
-                        key, redis_result, **kwargs
-                    )
-
-                result = redis_result
-
-            print_verbose(f"get cache: cache result: {result}")
-            return result
-        except Exception:
-            verbose_logger.error(traceback.format_exc())
-
-    async def async_batch_get_cache(
-        self, keys: list, local_only: bool = False, **kwargs
-    ):
-        try:
-            result = [None for _ in range(len(keys))]
-            if self.in_memory_cache is not None:
-                in_memory_result = await self.in_memory_cache.async_batch_get_cache(
-                    keys, **kwargs
-                )
-
-                if in_memory_result is not None:
-                    result = in_memory_result
-            if None in result and self.redis_cache is not None and local_only is False:
-                """
-                - for the none values in the result
-                - check the redis cache
-                """
-                sublist_keys = [
-                    key for key, value in zip(keys, result) if value is None
-                ]
-                # If not found in in-memory cache, try fetching from Redis
-                redis_result = await self.redis_cache.async_batch_get_cache(
-                    sublist_keys, **kwargs
-                )
-
-                if redis_result is not None:
-                    # Update in-memory cache with the value from Redis
-                    for key, value in redis_result.items():
-                        if value is not None:
-                            await self.in_memory_cache.async_set_cache(
-                                key, redis_result[key], **kwargs
-                            )
-                for key, value in redis_result.items():
-                    index = keys.index(key)
-                    result[index] = value
-
-            return result
-        except Exception:
-            verbose_logger.error(traceback.format_exc())
-
-    async def async_set_cache(self, key, value, local_only: bool = False, **kwargs):
-        print_verbose(
-            f"async set cache: cache key: {key}; local_only: {local_only}; value: {value}"
-        )
-        try:
-            if self.in_memory_cache is not None:
-                await self.in_memory_cache.async_set_cache(key, value, **kwargs)
-
-            if self.redis_cache is not None and local_only is False:
-                await self.redis_cache.async_set_cache(key, value, **kwargs)
-        except Exception as e:
-            verbose_logger.exception(
-                f"LiteLLM Cache: Excepton async add_cache: {str(e)}"
-            )
-
-    async def async_batch_set_cache(
-        self, cache_list: list, local_only: bool = False, **kwargs
-    ):
-        """
-        Batch write values to the cache
-        """
-        print_verbose(
-            f"async batch set cache: cache keys: {cache_list}; local_only: {local_only}"
-        )
-        try:
-            if self.in_memory_cache is not None:
-                await self.in_memory_cache.async_set_cache_pipeline(
-                    cache_list=cache_list, **kwargs
-                )
-
-            if self.redis_cache is not None and local_only is False:
-                await self.redis_cache.async_set_cache_pipeline(
-                    cache_list=cache_list, ttl=kwargs.pop("ttl", None), **kwargs
-                )
-        except Exception as e:
-            verbose_logger.exception(
-                f"LiteLLM Cache: Excepton async add_cache: {str(e)}"
-            )
-
-    async def async_increment_cache(
-        self, key, value: float, local_only: bool = False, **kwargs
-    ) -> float:
-        """
-        Key - the key in cache
-
-        Value - float - the value you want to increment by
-
-        Returns - float - the incremented value
-        """
-        try:
-            result: float = value
-            if self.in_memory_cache is not None:
-                result = await self.in_memory_cache.async_increment(
-                    key, value, **kwargs
-                )
-
-            if self.redis_cache is not None and local_only is False:
-                result = await self.redis_cache.async_increment(key, value, **kwargs)
-
-            return result
-        except Exception as e:
-            raise e  # don't log if exception is raised
-
-    async def async_set_cache_sadd(
-        self, key, value: List, local_only: bool = False, **kwargs
-    ) -> None:
-        """
-        Add value to a set
-
-        Key - the key in cache
-
-        Value - str - the value you want to add to the set
-
-        Returns - None
-        """
-        try:
-            if self.in_memory_cache is not None:
-                _ = await self.in_memory_cache.async_set_cache_sadd(
-                    key, value, ttl=kwargs.get("ttl", None)
-                )
-
-            if self.redis_cache is not None and local_only is False:
-                _ = await self.redis_cache.async_set_cache_sadd(
-                    key, value, ttl=kwargs.get("ttl", None) ** kwargs
-                )
-
-            return None
-        except Exception as e:
-            raise e  # don't log, if exception is raised
-
-    def flush_cache(self):
-        if self.in_memory_cache is not None:
-            self.in_memory_cache.flush_cache()
-        if self.redis_cache is not None:
-            self.redis_cache.flush_cache()
-
-    def delete_cache(self, key):
-        """
-        Delete a key from the cache
-        """
-        if self.in_memory_cache is not None:
-            self.in_memory_cache.delete_cache(key)
-        if self.redis_cache is not None:
-            self.redis_cache.delete_cache(key)
-
-    async def async_delete_cache(self, key: str):
-        """
-        Delete a key from the cache
-        """
-        if self.in_memory_cache is not None:
-            self.in_memory_cache.delete_cache(key)
-        if self.redis_cache is not None:
-            await self.redis_cache.async_delete_cache(key)
 
 
 #### LiteLLM.Completion / Embedding Cache ####
@@ -1285,84 +632,6 @@ class Cache:
         if self.type and self.type == LiteLLMCacheType.S3:
             return False
         return True
-
-
-class DiskCache(BaseCache):
-    def __init__(self, disk_cache_dir: Optional[str] = None):
-        import diskcache as dc
-
-        # if users don't provider one, use the default litellm cache
-        if disk_cache_dir is None:
-            self.disk_cache = dc.Cache(".litellm_cache")
-        else:
-            self.disk_cache = dc.Cache(disk_cache_dir)
-
-    def set_cache(self, key, value, **kwargs):
-        print_verbose("DiskCache: set_cache")
-        if "ttl" in kwargs:
-            self.disk_cache.set(key, value, expire=kwargs["ttl"])
-        else:
-            self.disk_cache.set(key, value)
-
-    async def async_set_cache(self, key, value, **kwargs):
-        self.set_cache(key=key, value=value, **kwargs)
-
-    async def async_set_cache_pipeline(self, cache_list, ttl=None):
-        for cache_key, cache_value in cache_list:
-            if ttl is not None:
-                self.set_cache(key=cache_key, value=cache_value, ttl=ttl)
-            else:
-                self.set_cache(key=cache_key, value=cache_value)
-
-    def get_cache(self, key, **kwargs):
-        original_cached_response = self.disk_cache.get(key)
-        if original_cached_response:
-            try:
-                cached_response = json.loads(original_cached_response)  # type: ignore
-            except Exception:
-                cached_response = original_cached_response
-            return cached_response
-        return None
-
-    def batch_get_cache(self, keys: list, **kwargs):
-        return_val = []
-        for k in keys:
-            val = self.get_cache(key=k, **kwargs)
-            return_val.append(val)
-        return return_val
-
-    def increment_cache(self, key, value: int, **kwargs) -> int:
-        # get the value
-        init_value = self.get_cache(key=key) or 0
-        value = init_value + value  # type: ignore
-        self.set_cache(key, value, **kwargs)
-        return value
-
-    async def async_get_cache(self, key, **kwargs):
-        return self.get_cache(key=key, **kwargs)
-
-    async def async_batch_get_cache(self, keys: list, **kwargs):
-        return_val = []
-        for k in keys:
-            val = self.get_cache(key=k, **kwargs)
-            return_val.append(val)
-        return return_val
-
-    async def async_increment(self, key, value: int, **kwargs) -> int:
-        # get the value
-        init_value = await self.async_get_cache(key=key) or 0
-        value = init_value + value  # type: ignore
-        await self.async_set_cache(key, value, **kwargs)
-        return value
-
-    def flush_cache(self):
-        self.disk_cache.clear()
-
-    async def disconnect(self):
-        pass
-
-    def delete_cache(self, key):
-        self.disk_cache.pop(key)
 
 
 def enable_cache(
